+++ .eslintrc.js
@@ -1,5 +1,7 @@
 module.exports = {
   extends: [
+    'plugin:eslint-comments/recommended',
+
     'standard',
     'standard-jsx',
     'prettier',
@@ -19,7 +21,7 @@ module.exports = {
 
   overrides: [
     {
-      files: ['packages/*cli*/**/*.js', '*-cli.js'],
+      files: ['cli.js', '*-cli.js', 'packages/*cli*/**/*.js'],
       rules: {
         'no-console': 'off',
       },