+++ packages/xo-server/src/xapi/mixins/vm.js
@@ -52,6 +52,7 @@ export default {
 
       coreOs = false,
       cloudConfig = undefined,
+      networkConfig = undefined,
 
       vgpuType = undefined,
       gpuGroup = undefined,
@@ -241,10 +242,16 @@ export default {
         }
       })
 
-      const method = coreOs
-        ? 'createCoreOsCloudInitConfigDrive'
-        : 'createCloudInitConfigDrive'
-      await this[method](vm.$id, srRef, cloudConfig)
+      if (coreOs) {
+        await this.createCoreOsCloudInitConfigDrive(vm.$id, srRef, cloudConfig)
+      } else {
+        await this.createCloudInitConfigDrive(
+          vm.$id,
+          srRef,
+          cloudConfig,
+          networkConfig
+        )
+      }
     }
 
     // wait for the record with all the VBDs and VIFs