+++ packages/xo-server/src/xo-mixins/jobs/index.js
@@ -164,11 +164,11 @@ export default class Jobs {
 
         xo.emit(
           'job:terminated',
-          undefined,
-          job,
-          undefined,
           // This cast can be removed after merging the PR: https://github.com/vatesfr/xen-orchestra/pull/3209
-          String(job.runId)
+          String(job.runId),
+          {
+            type: job.type,
+          }
         )
         return this.updateJob({ id: job.id, runId: null })
       })
@@ -266,6 +266,11 @@ export default class Jobs {
         reportWhen: (settings && settings.reportWhen) || 'failure',
       }
     }
+    if (type === 'metadataBackup') {
+      data = {
+        reportWhen: job.settings['']?.reportWhen ?? 'failure',
+      }
+    }
 
     const logger = this._logger
     const runJobId = logger.notice(`Starting execution of ${id}.`, {
@@ -314,7 +319,10 @@ export default class Jobs {
         true
       )
 
-      app.emit('job:terminated', status, job, schedule, runJobId)
+      app.emit('job:terminated', runJobId, {
+        type: job.type,
+        status,
+      })
     } catch (error) {
       await logger.error(
         `The execution of ${id} has failed.`,
@@ -325,7 +333,9 @@ export default class Jobs {
         },
         true
       )
-      app.emit('job:terminated', undefined, job, schedule, runJobId)
+      app.emit('job:terminated', runJobId, {
+        type: job.type,
+      })
       throw error
     } finally {
       this.updateJob({ id, runId: null })::ignoreErrors()