+++ packages/xo-web/src/common/cloud-config.js
@@ -21,6 +21,37 @@ const showAvailableTemplateVars = () =>
     </ul>
   )
 
+const showNetworkConfigInfo = () =>
+  alert(
+    _('newVmNetworkConfigLabel'),
+    <div>
+      <p>
+        {_('newVmNetworkConfigInfo', {
+          noCloudDatasourceLink: (
+            <a
+              href='https://cloudinit.readthedocs.io/en/latest/topics/datasources/nocloud.html#datasource-nocloud'
+              target='_blank'
+            >
+              {_('newVmNoCloudDatasource')}
+            </a>
+          ),
+        })}
+      </p>
+      <p>
+        {_('newVmNetworkConfigDocLink', {
+          networkConfigDocLink: (
+            <a
+              href='https://cloudinit.readthedocs.io/en/latest/topics/network-config-format-v1.html'
+              target='_blank'
+            >
+              {_('newVmNetworkConfigDoc')}
+            </a>
+          ),
+        })}
+      </p>
+    </div>
+  )
+
 export const AvailableTemplateVars = () => (
   <Tooltip content={_('availableTemplateVarsInfo')}>
     <a
@@ -33,5 +64,26 @@ export const AvailableTemplateVars = () => (
   </Tooltip>
 )
 
+export const NetworkConfigInfo = () => (
+  <Tooltip content={_('newVmNetworkConfigTooltip')}>
+    <a
+      className='text-info'
+      style={{ cursor: 'pointer' }}
+      onClick={showNetworkConfigInfo}
+    >
+      <Icon icon='info' />
+    </a>
+  </Tooltip>
+)
+
 export const DEFAULT_CLOUD_CONFIG_TEMPLATE =
   '#cloud-config\n#hostname: {name}%\n#ssh_authorized_keys:\n#  - ssh-rsa <myKey>\n#packages:\n#  - htop\n'
+
+// SOURCE: https://cloudinit.readthedocs.io/en/latest/topics/network-config-format-v1.html
+export const DEFAULT_NETWORK_CONFIG_TEMPLATE = `#network:
+#  version: 1
+#  config:
+#  - type: physical
+#    name: eth0
+#    subnets:
+#      - type: dhcp`