+++ packages/xo-web/src/common/intl/locales/fr.js
@@ -3016,7 +3016,8 @@ export default {
     "Le SR par défaut n'est pas connecté à l'hôte",
 
   // Original text: "For each VDI, select an SR:"
-  chooseSrForEachVdisModalSelectSr: 'Pour chaque VDI, sélectionner un SR :',
+  chooseSrForEachVdisModalSelectSr:
+    'Pour chaque VDI, sélectionner un SR (optionnel)',
 
   // Original text: "Select main SR…"
   chooseSrForEachVdisModalMainSr: 'Sélectionner le SR principal…',
@@ -3496,7 +3497,7 @@ export default {
 
   // Original text: "This version is not bundled with any support nor updates. Use it with caution for critical tasks."
   disclaimerText3:
-    "Cette version n'est fournie avec aucun support ni aucune mise à jour. Soyez prudent en cas d'utilisation pour des tâches importantes.",
+    "Cette version n'est fournie avec aucun support ni aucune mise à jour. Utilisez-la avec précaution.",
 
   // Original text: "Connect PIF"
   connectPif: 'Connecter la PIF',