+++ packages/xo-web/src/xo-app/new-vm/index.js
@@ -17,7 +17,10 @@ import Wizard, { Section } from 'wizard'
 import {
   AvailableTemplateVars,
   DEFAULT_CLOUD_CONFIG_TEMPLATE,
+  DEFAULT_NETWORK_CONFIG_TEMPLATE,
+  NetworkConfigInfo,
 } from 'cloud-config'
+import { confirm } from 'modal'
 import { Container, Row, Col } from 'grid'
 import { injectIntl } from 'react-intl'
 import {
@@ -79,19 +82,21 @@ import {
   getCoresPerSocketPossibilities,
   generateReadableRandomString,
   resolveIds,
-  resolveResourceSet,
 } from 'utils'
 import {
   createFilter,
-  createSelector,
+  createFinder,
   createGetObject,
   createGetObjectsOfType,
+  createSelector,
   getIsPoolAdmin,
+  getResolvedResourceSets,
   getUser,
 } from 'selectors'
 
 import styles from './index.css'
 
+const MULTIPLICAND = 2
 const NB_VMS_MIN = 2
 const NB_VMS_MAX = 100
 
@@ -221,25 +226,39 @@ class Vif extends BaseComponent {
   resourceSets: subscribeResourceSets,
   user: subscribeCurrentUser,
 })
-@connectStore(() => ({
-  isAdmin: createSelector(
+@connectStore(() => {
+  const getIsAdmin = createSelector(
     getUser,
     user => user && user.permission === 'admin'
-  ),
-  isPoolAdmin: getIsPoolAdmin,
-  networks: createGetObjectsOfType('network').sort(),
-  pool: createGetObject((_, props) => props.location.query.pool),
-  pools: createGetObjectsOfType('pool'),
-  templates: createGetObjectsOfType('VM-template').sort(),
-  userSshKeys: createSelector(
+  )
+  const getNetworks = createGetObjectsOfType('network').sort()
+  const getPool = createGetObject((_, props) => props.location.query.pool)
+  const getPools = createGetObjectsOfType('pool')
+  const getSrs = createGetObjectsOfType('SR')
+  const getTemplates = createGetObjectsOfType('VM-template').sort()
+  const getUserSshKeys = createSelector(
     (_, props) => {
       const user = props.user
       return user && user.preferences && user.preferences.sshKeys
     },
     keys => keys
-  ),
-  srs: createGetObjectsOfType('SR'),
-}))
+  )
+  return (state, props) => ({
+    isAdmin: getIsAdmin(state, props),
+    isPoolAdmin: getIsPoolAdmin(state, props),
+    networks: getNetworks(state, props),
+    pool: getPool(state, props),
+    pools: getPools(state, props),
+    resolvedResourceSets: getResolvedResourceSets(
+      state,
+      props,
+      props.pool === undefined // to get objects as a self user
+    ),
+    srs: getSrs(state, props),
+    templates: getTemplates(state, props),
+    userSshKeys: getUserSshKeys(state, props),
+  })
+})
 @injectIntl
 export default class NewVm extends BaseComponent {
   static contextTypes = {
@@ -259,19 +278,24 @@ export default class NewVm extends BaseComponent {
     this._reset()
   }
 
-  _getResourceSet = () => {
-    const {
-      location: {
-        query: { resourceSet: resourceSetId },
-      },
-      resourceSets,
-    } = this.props
-    return resourceSets && find(resourceSets, ({ id }) => id === resourceSetId)
-  }
+  _getResourceSet = createFinder(
+    () => this.props.resourceSets,
+    createSelector(
+      () => this.props.location.query.resourceSet,
+      resourceSetId => resourceSet =>
+        resourceSet !== undefined ? resourceSetId === resourceSet.id : undefined
+    )
+  )
 
-  _getResolvedResourceSet = createSelector(
-    this._getResourceSet,
-    resolveResourceSet
+  _getResolvedResourceSet = createFinder(
+    () => this.props.resolvedResourceSets,
+    createSelector(
+      this._getResourceSet,
+      resourceSet =>
+        resourceSet !== undefined
+          ? resolvedResourceSet => resolvedResourceSet.id === resourceSet.id
+          : false
+    )
   )
 
   // Utils -----------------------------------------------------------------------
@@ -324,6 +348,29 @@ export default class NewVm extends BaseComponent {
     })
   }
 
+  _selfCreate = () => {
+    const {
+      CPUs,
+      VDIs,
+      existingDisks,
+      memoryDynamicMax,
+      template,
+    } = this.state.state
+    const disksSize = sumBy(VDIs, 'size') + sumBy(existingDisks, 'size')
+    const templateDisksSize = sumBy(template.template_info.disks, 'size')
+    const templateMemoryDynamicMax = template.memory.dynamic[1]
+    const templateVcpusMax = template.CPUs.max
+
+    return CPUs > MULTIPLICAND * templateVcpusMax ||
+      memoryDynamicMax > MULTIPLICAND * templateMemoryDynamicMax ||
+      disksSize > MULTIPLICAND * templateDisksSize
+      ? confirm({
+          title: _('createVmModalTitle'),
+          body: _('createVmModalWarningMessage'),
+        }).then(this._create)
+      : this._create()
+  }
+
   _create = () => {
     const { state } = this.state
     let installation
@@ -353,6 +400,7 @@ export default class NewVm extends BaseComponent {
 
     let cloudConfig
     let cloudConfigs
+    let networkConfig
     if (state.installMethod !== 'noConfigDrive') {
       if (state.installMethod === 'SSH') {
         const format = hostname =>
@@ -389,8 +437,12 @@ export default class NewVm extends BaseComponent {
             replacer(state, i + +seqStart)
           )
         }
+        networkConfig = defined(
+          state.networkConfig,
+          DEFAULT_NETWORK_CONFIG_TEMPLATE
+        )
       }
-    } else if (state.template.name_label === 'CoreOS') {
+    } else if (this._isCoreOs()) {
       cloudConfig = state.cloudConfig
       if (state.multipleVms) {
         cloudConfigs = new Array(state.nbVms).fill(state.cloudConfig)
@@ -443,7 +495,8 @@ export default class NewVm extends BaseComponent {
       bootAfterCreate: state.bootAfterCreate,
       share: state.share,
       cloudConfig,
-      coreOs: state.template.name_label === 'CoreOS',
+      networkConfig: this._isCoreOs() ? undefined : networkConfig,
+      coreOs: this._isCoreOs(),
       tags: state.tags,
       vgpuType: get(() => state.vgpuType.id),
       gpuGroup: get(() => state.vgpuType.gpuGroup),
@@ -543,7 +596,7 @@ export default class NewVm extends BaseComponent {
       }),
     })
 
-    if (template.name_label === 'CoreOS') {
+    if (this._isCoreOs()) {
       getCloudInitConfig(template.id).then(
         cloudConfig =>
           this._setState({ cloudConfig, coreOsDefaultTemplateError: false }),
@@ -691,6 +744,11 @@ export default class NewVm extends BaseComponent {
     },
     () => this.state.state.CPUs,
     getCoresPerSocketPossibilities
+  )
+
+  _isCoreOs = createSelector(
+    () => this.state.template,
+    template => template && template.name_label === 'CoreOS'
   )
 
   // On change -------------------------------------------------------------------
@@ -915,7 +973,7 @@ export default class NewVm extends BaseComponent {
                   ) || !this._availableResources()
                 }
                 form='vmCreation'
-                handler={this._create}
+                handler={pool === undefined ? this._selfCreate : this._create}
                 icon='new-vm-create'
                 redirectOnSuccess={this._getRedirectionUrl}
               >
@@ -1072,6 +1130,7 @@ export default class NewVm extends BaseComponent {
     const {
       cloudConfig,
       customConfig,
+      networkConfig,
       installIso,
       installMethod,
       installNetwork,
@@ -1163,13 +1222,39 @@ export default class NewVm extends BaseComponent {
               </span>
             </LineItem>
             <br />
-            <DebounceTextarea
-              className='form-control'
-              disabled={installMethod !== 'customConfig'}
-              onChange={this._linkState('customConfig')}
-              rows={7}
-              value={defined(customConfig, DEFAULT_CLOUD_CONFIG_TEMPLATE)}
-            />
+            <LineItem>
+              <Item>
+                <label className='text-muted'>
+                  {_('newVmUserConfigLabel')}
+                  <br />
+                  <DebounceTextarea
+                    className='form-control'
+                    disabled={installMethod !== 'customConfig'}
+                    onChange={this._linkState('customConfig')}
+                    rows={7}
+                    value={defined(customConfig, DEFAULT_CLOUD_CONFIG_TEMPLATE)}
+                  />
+                </label>
+              </Item>
+              {!this._isCoreOs() && (
+                <Item>
+                  <label className='text-muted'>
+                    {_('newVmNetworkConfigLabel')} <NetworkConfigInfo />
+                    <br />
+                    <DebounceTextarea
+                      className='form-control'
+                      disabled={installMethod !== 'customConfig'}
+                      onChange={this._linkState('networkConfig')}
+                      rows={7}
+                      value={defined(
+                        networkConfig,
+                        DEFAULT_NETWORK_CONFIG_TEMPLATE
+                      )}
+                    />
+                  </label>
+                </Item>
+              )}
+            </LineItem>
           </SectionContent>
         ) : (
           <SectionContent>
@@ -1249,7 +1334,7 @@ export default class NewVm extends BaseComponent {
             )}
           </SectionContent>
         )}
-        {template.name_label === 'CoreOS' && (
+        {this._isCoreOs() && (
           <div>
             <label>{_('newVmCloudConfig')}</label>{' '}
             {!coreOsDefaultTemplateError ? (